+++ core/src/DataColumn.hpp
@@ -56,7 +56,7 @@ namespace models::stats {
   }
 
   template<typename T>
-  double accuracy(const DataColumn<T> &predictions, const DataColumn<T> &actual) {
+  float accuracy(const DataColumn<T> &predictions, const DataColumn<T> &actual) {
     if (predictions.rows() != actual.rows()) {
       throw std::invalid_argument("predictions and actual must have the same number of rows");
     }
@@ -69,11 +69,11 @@ namespace models::stats {
       }
     }
 
-    return (double)correct / (double)predictions.rows();
+    return (float)correct / (float)predictions.rows();
   }
 
   template<typename T>
-  double error_rate(const DataColumn<T> &predictions, const DataColumn<T> &actual) {
+  float error_rate(const DataColumn<T> &predictions, const DataColumn<T> &actual) {
     if (predictions.rows() != actual.rows()) {
       throw std::invalid_argument("predictions and actual must have the same number of rows");
     }
