+++ core/src/Math.hpp
@@ -5,15 +5,15 @@
 #include "Macros.hpp"
 
 namespace models::math {
-  inline bool is_approx(double a, double b, double threshold) {
+  inline bool is_approx(float a, float b, float threshold) {
     return fabs(a - b) < threshold;
   }
 
-  inline bool is_approx(double a, double b) {
+  inline bool is_approx(float a, float b) {
     return is_approx(a, b, APPROX_THRESHOLD);
   }
 
-  inline bool is_module_approx(double a, double b) {
+  inline bool is_module_approx(float a, float b) {
     return is_approx(fabs(a), fabs(b));
   }
 }
